requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
langchain_openai==0.0.2
python-dotenv==1.0.0
scapy==2.5.0
reportlab==4.0.4
matplotlib==3.7.3
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.4
fpdf==1.7.2
pyshark==0.6

